+++ languages/en/tl_news_feed.php
@@ -15,4 +15,13 @@
 /**
  * Fields
  */
-$GLOBALS['TL_LANG']['tl_news_feed']['categories'] = array('Categories', 'Here you can select one or more news categories.');
+$GLOBALS['TL_LANG']['tl_news_feed']['categories']      = array('Categories', 'Here you can select one or more news categories.');
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show'] = array('Show categories in the feed', 'Here you can choose how the news categories will be showed in the feed.');
+
+/**
+ * Reference
+ */
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['empty']       = 'Do not show';
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['title']       = 'Show in title';
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['text_before'] = 'Show before the text';
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['text_after']  = 'Show after the text';