+++ languages/pl/default.php
@@ -15,4 +15,10 @@
 /**
  * Miscellaneous
  */
+$GLOBALS['TL_LANG']['MSC']['newsCategories']  = 'Kategorie:';
 $GLOBALS['TL_LANG']['MSC']['resetCategories'] = array('Wszystkie', 'Pokaż aktualności z wszystkich kategorii');
+
+/**
+ * Content elements
+ */
+$GLOBALS['TL_LANG']['CTE']['newsfilter'] = array('Moduł aktualności filtrowany', 'Dodaje moduł aktualności do strony z własnymi ustawieniami filtrowania.');