+++ languages/pl/tl_news_feed.php
@@ -15,4 +15,13 @@
 /**
  * Fields
  */
-$GLOBALS['TL_LANG']['tl_news_feed']['categories'] = array('Kategorie', 'Tutaj możesz wybrać jedną lub więcej kategorii.');
+$GLOBALS['TL_LANG']['tl_news_feed']['categories']      = array('Kategorie', 'Tutaj możesz wybrać jedną lub więcej kategorii.');
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show'] = array('Pokaż kategorie w kanale', 'Tutaj możesz wybrać w jaki sposób kategorie zostaną pokazane w kanale.');
+
+/**
+ * Reference
+ */
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['empty']       = 'Nie pokazuj';
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['title']       = 'Pokaż w tytule';
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['text_before'] = 'Pokaż przed tekstem';
+$GLOBALS['TL_LANG']['tl_news_feed']['categories_show']['text_after']  = 'Pokaż po tekście';