+++ modules/ModuleNewsMenu.php
@@ -27,9 +27,7 @@ class ModuleNewsMenu extends \Contao\ModuleNewsMenu
      */
     public function generate()
     {
-        $GLOBALS['NEWS_FILTER_CATEGORIES'] = $this->news_filterCategories ? true : false;
-        $GLOBALS['NEWS_FILTER_DEFAULT'] = deserialize($this->news_filterDefault, true);
-        $GLOBALS['NEWS_FILTER_PRESERVE'] = $this->news_filterPreserve;
+        $this->news_filterDefault = deserialize($this->news_filterDefault, true);
 
         return parent::generate();
     }
@@ -39,11 +37,44 @@ class ModuleNewsMenu extends \Contao\ModuleNewsMenu
      */
     protected function compileYearlyMenu()
     {
-        parent::compileYearlyMenu();
+        $time = time();
+        $arrData = array();
+        $this->Template = new \FrontendTemplate('mod_newsmenu_year');
+        $arrNewsIds = $this->getFilteredNewsIds();
 
-        if ($this->news_filterCategories) {
-            $this->updateMenuLinks();
+        // Get the dates
+        $objDates = $this->Database->query("SELECT FROM_UNIXTIME(date, '%Y') AS year, COUNT(*) AS count FROM tl_news WHERE pid IN(" . implode(',', array_map('intval', $this->news_archives)) . ")" . ((!BE_USER_LOGGED_IN || TL_MODE == 'BE') ? " AND (start='' OR start<$time) AND (stop='' OR stop>$time) AND published=1" : "") . (!empty($arrNewsIds) ? (" AND id IN (" . implode(',', $arrNewsIds) . ")") : "") . " GROUP BY year ORDER BY year DESC");
+
+        while ($objDates->next())
+        {
+            $arrData[$objDates->year] = $objDates->count;
         }
+
+        // Sort the data
+        ($this->news_order == 'ascending') ? ksort($arrData) : krsort($arrData);
+
+        $arrItems = array();
+        $count = 0;
+        $limit = count($arrData);
+        $strUrl = $this->generateCategoryUrl();
+
+        // Prepare the navigation
+        foreach ($arrData as $intYear=>$intCount)
+        {
+            $intDate = $intYear;
+            $quantity = sprintf((($intCount < 2) ? $GLOBALS['TL_LANG']['MSC']['entry'] : $GLOBALS['TL_LANG']['MSC']['entries']), $intCount);
+
+            $arrItems[$intYear]['date'] = $intDate;
+            $arrItems[$intYear]['link'] = $intYear;
+            $arrItems[$intYear]['href'] = $strUrl . ($GLOBALS['TL_CONFIG']['disableAlias'] ? '&amp;' : '?') . 'year=' . $intDate;
+            $arrItems[$intYear]['title'] = specialchars($intYear . ' (' . $quantity . ')');
+            $arrItems[$intYear]['class'] = trim(((++$count == 1) ? 'first ' : '') . (($count == $limit) ? 'last' : ''));
+            $arrItems[$intYear]['isActive'] = (\Input::get('year') == $intDate);
+            $arrItems[$intYear]['quantity'] = $quantity;
+        }
+
+        $this->Template->items = $arrItems;
+        $this->Template->showQuantity = ($this->news_showQuantity != '');
     }
 
     /**
@@ -51,35 +82,56 @@ class ModuleNewsMenu extends \Contao\ModuleNewsMenu
      */
     protected function compileMonthlyMenu()
     {
-        parent::compileMonthlyMenu();
+        $time = time();
+        $arrData = array();
+        $arrNewsIds = $this->getFilteredNewsIds();
 
-        if ($this->news_filterCategories) {
-            $this->updateMenuLinks();
+        // Get the dates
+        $objDates = $this->Database->query("SELECT FROM_UNIXTIME(date, '%Y') AS year, FROM_UNIXTIME(date, '%m') AS month, COUNT(*) AS count FROM tl_news WHERE pid IN(" . implode(',', array_map('intval', $this->news_archives)) . ")" . ((!BE_USER_LOGGED_IN || TL_MODE == 'BE') ? " AND (start='' OR start<$time) AND (stop='' OR stop>$time) AND published=1" : "") . (!empty($arrNewsIds) ? (" AND id IN (" . implode(',', $arrNewsIds) . ")") : "") . " GROUP BY year, month ORDER BY year DESC, month DESC");
+
+        while ($objDates->next())
+        {
+            $arrData[$objDates->year][$objDates->month] = $objDates->count;
         }
-    }
 
-    /**
-     * Update the menu links by adding a category
-     */
-    protected function updateMenuLinks()
-    {
+        // Sort the data
+        foreach (array_keys($arrData) as $key)
+        {
+            ($this->news_order == 'ascending') ? ksort($arrData[$key]) : krsort($arrData[$key]);
+        }
+
+        ($this->news_order == 'ascending') ? ksort($arrData) : krsort($arrData);
+
         $strUrl = $this->generateCategoryUrl();
-        $arrItems = $this->Template->items;
+        $arrItems = array();
 
-        // Update the links
-        foreach ($arrItems as $k => $v) {
-            if (isset($v['href'])) {
-                $params = explode('?', $v['href']);
-                $arrItems[$k]['href'] = $strUrl . '?' . $params[1];
-            } else {
-                foreach ($v as $kk => $vv) {
-                    $params = explode('?', $vv['href']);
-                    $arrItems[$k][$kk]['href'] = $strUrl . '?' . $params[1];
-                }
+        // Prepare the navigation
+        foreach ($arrData as $intYear=>$arrMonth)
+        {
+            $count = 0;
+            $limit = count($arrMonth);
+
+            foreach ($arrMonth as $intMonth=>$intCount)
+            {
+                $intDate = $intYear . $intMonth;
+                $intMonth = (intval($intMonth) - 1);
+
+                $quantity = sprintf((($intCount < 2) ? $GLOBALS['TL_LANG']['MSC']['entry'] : $GLOBALS['TL_LANG']['MSC']['entries']), $intCount);
+
+                $arrItems[$intYear][$intMonth]['date'] = $intDate;
+                $arrItems[$intYear][$intMonth]['link'] = $GLOBALS['TL_LANG']['MONTHS'][$intMonth] . ' ' . $intYear;
+                $arrItems[$intYear][$intMonth]['href'] = $strUrl . ($GLOBALS['TL_CONFIG']['disableAlias'] ? '&amp;' : '?') . 'month=' . $intDate;
+                $arrItems[$intYear][$intMonth]['title'] = specialchars($GLOBALS['TL_LANG']['MONTHS'][$intMonth].' '.$intYear . ' (' . $quantity . ')');
+                $arrItems[$intYear][$intMonth]['class'] = trim(((++$count == 1) ? 'first ' : '') . (($count == $limit) ? 'last' : ''));
+                $arrItems[$intYear][$intMonth]['isActive'] = (\Input::get('month') == $intDate);
+                $arrItems[$intYear][$intMonth]['quantity'] = $quantity;
             }
         }
 
         $this->Template->items = $arrItems;
+        $this->Template->showQuantity = ($this->news_showQuantity != '') ? true : false;
+        $this->Template->url = $strUrl . ($GLOBALS['TL_CONFIG']['disableAlias'] ? '&amp;' : '?');
+        $this->Template->activeYear = \Input::get('year');
     }
 
     /**
@@ -87,15 +139,64 @@ class ModuleNewsMenu extends \Contao\ModuleNewsMenu
      */
     protected function compileDailyMenu()
     {
-        parent::compileDailyMenu();
+        $time = time();
+        $arrData = array();
+        $this->Template = new \FrontendTemplate('mod_newsmenu_day');
+        $arrNewsIds = $this->getFilteredNewsIds();
 
-        if ($this->news_filterCategories) {
-            $prevParams = explode('?', $this->Template->prevHref);
-            $this->Template->prevHref = $this->generateCategoryUrl() . '?' . $prevParams[1];
+        // Get the dates
+        $objDates = $this->Database->query("SELECT FROM_UNIXTIME(date, '%Y%m%d') AS day, COUNT(*) AS count FROM tl_news WHERE pid IN(" . implode(',', array_map('intval', $this->news_archives)) . ")" . ((!BE_USER_LOGGED_IN || TL_MODE == 'BE') ? " AND (start='' OR start<$time) AND (stop='' OR stop>$time) AND published=1" : "") . (!empty($arrNewsIds) ? (" AND id IN (" . implode(',', $arrNewsIds) . ")") : "") . " GROUP BY day ORDER BY day DESC");
 
-            $nextParams = explode('?', $this->Template->nextHref);
-            $this->Template->nextHref = $this->generateCategoryUrl() . '?' . $nextParams[1];
+        while ($objDates->next())
+        {
+            $arrData[$objDates->day] = $objDates->count;
         }
+
+        // Sort the data
+        krsort($arrData);
+        $strUrl = $this->generateCategoryUrl();
+
+        $this->Date = \Input::get('day') ? new \Date(\Input::get('day'), 'Ymd') : new \Date();
+
+        $intYear = date('Y', $this->Date->tstamp);
+        $intMonth = date('m', $this->Date->tstamp);
+
+        $this->Template->intYear = $intYear;
+        $this->Template->intMonth = $intMonth;
+
+        // Previous month
+        $prevMonth = ($intMonth == 1) ? 12 : ($intMonth - 1);
+        $prevYear = ($intMonth == 1) ? ($intYear - 1) : $intYear;
+        $lblPrevious = $GLOBALS['TL_LANG']['MONTHS'][($prevMonth - 1)] . ' ' . $prevYear;
+
+        $this->Template->prevHref = $strUrl . ($GLOBALS['TL_CONFIG']['disableAlias'] ? '?id=' . \Input::get('id') . '&amp;' : '?') . 'day=' . $prevYear . ((strlen($prevMonth) < 2) ? '0' : '') . $prevMonth . '01';
+        $this->Template->prevTitle = specialchars($lblPrevious);
+        $this->Template->prevLink = $GLOBALS['TL_LANG']['MSC']['news_previous'] . ' ' . $lblPrevious;
+        $this->Template->prevLabel = $GLOBALS['TL_LANG']['MSC']['news_previous'];
+
+        // Current month
+        $this->Template->current = $GLOBALS['TL_LANG']['MONTHS'][(date('m', $this->Date->tstamp) - 1)] .  ' ' . date('Y', $this->Date->tstamp);
+
+        // Next month
+        $nextMonth = ($intMonth == 12) ? 1 : ($intMonth + 1);
+        $nextYear = ($intMonth == 12) ? ($intYear + 1) : $intYear;
+        $lblNext = $GLOBALS['TL_LANG']['MONTHS'][($nextMonth - 1)] . ' ' . $nextYear;
+
+        $this->Template->nextHref = $strUrl . ($GLOBALS['TL_CONFIG']['disableAlias'] ? '?id=' . \Input::get('id') . '&amp;' : '?') . 'day=' . $nextYear . ((strlen($nextMonth) < 2) ? '0' : '') . $nextMonth . '01';
+        $this->Template->nextTitle = specialchars($lblNext);
+        $this->Template->nextLink = $lblNext . ' ' . $GLOBALS['TL_LANG']['MSC']['news_next'];
+        $this->Template->nextLabel = $GLOBALS['TL_LANG']['MSC']['news_next'];
+
+        // Set week start day
+        if (!$this->news_startDay)
+        {
+            $this->news_startDay = 0;
+        }
+
+        $this->Template->days = $this->compileDays();
+        $this->Template->weeks = $this->compileWeeks($arrData, $strUrl);
+
+        $this->Template->showQuantity = ($this->news_showQuantity != '') ? true : false;
     }
 
     /**
@@ -115,7 +216,7 @@ class ModuleNewsMenu extends \Contao\ModuleNewsMenu
      */
     protected function generateCategoryUrl()
     {
-        $strUrl = '';
+        $strUrl = \Environment::get('request');
 
         // Get the current "jumpTo" page
         if (($objTarget = $this->objModel->getRelated('jumpTo')) !== null) {
@@ -130,5 +231,49 @@ class ModuleNewsMenu extends \Contao\ModuleNewsMenu
         }
 
         return $strUrl;
+    }
+
+    /**
+     * Get the filtered news IDs
+     *
+     * @return array
+     */
+    protected function getFilteredNewsIds()
+    {
+        $arrCategories = \NewsCategories\NewsModel::getCategoriesCache();
+
+        if (empty($arrCategories)) {
+            return array();
+        }
+
+        $arrIds = array();
+
+        // Use the default filter
+        if (is_array($this->news_filterDefault) && !empty($this->news_filterDefault)) {
+            foreach ($this->news_filterDefault as $category) {
+                if (isset($arrCategories[$category])) {
+                    $arrIds = array_merge($arrCategories[$category], $arrIds);
+                }
+            }
+        }
+
+        // Current category
+        if ($this->news_filterCategories && \Input::get('category')) {
+            $strClass = \NewsCategories\NewsCategories::getModelClass();
+            $objCategory = $strClass::findPublishedByIdOrAlias(\Input::get('category'));
+
+            if ($objCategory === null) {
+                return array();
+            }
+
+            // Preserve the default filter
+            if ($this->news_filterPreserve) {
+                $arrIds = array_merge($arrCategories[$objCategory->id], $arrIds);
+            } else {
+                $arrIds = $arrCategories[$objCategory->id];
+            }
+        }
+
+        return array_unique($arrIds);
     }
 }