+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 news_categories Changelog
 =========================
 
+Version 2.2.0 stable (2015-01-23)
+---------------------------------
+
+### Improved
+- Made the news menu filterable by categories (see #46)
+- Added the composer.json file (see #34)
+- Added the categories to news feed (see #47)
+- Added the news filter content element (see #51)
+- Added the dc_multlingual to composer.json file (see #50)
+
+
 Version 2.1.2 stable (2014-11-03)
 ---------------------------------